Use helper to try multiple key paths for user key
Ticket: BG-24941
